MyBot/requirements.txt
discord.py==2.3.2
python_dotenv
load_dotenv
flask
pytz
requests
beautifulsoup4
aiohttp
openai
responses
feedparser
discord.py
geopy
timezonefinder
googletrans==4.0.0-rc1
requests
pytz